Remove rgeos dependency and fix examples
The rgeos package is not on CRAN anymore.  Also, it does not seem to
be used by this package at the moment.

Two examples that were not valid were commented out.  A TODO note was
added as a reminder that they might be written in the future.

man/hello.Rd was deleted.